+++ CHANGELOG.md
@@ -6,3 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## Added
+
+- Chemical Storage Disk
+- Chemical support in the Grid.
+- Chemical support in the Storage Monitor.